Add random RGB color change for each square
This is the first task of the extra credit for the Etch-a-Sketch project. Instead of a simple color change from black to white, each interaction should randomize the square's RGB value entirely. I modified the function that changes the color of the div to generate and apply a random RGB color.